+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## `jupyter_starters 0.1.0a2`
+
+- add ignore patterns to schema
+- fix default ignore patterns
+
+## `@deathbeds/jupyterlab-starters 0.1.0a2`
+
+- add glob ignore patterns to schema
+
 ## `jupyter_starters 0.1.0a1`
 
 - add more sources of config